# backend/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CashX API
  version: 1.0.0
  description: Partner platform API. All money in kopecks (int64), rates in basis points.
servers:
  - url: /api/v1
paths:
...
  /cabinet/offers/{offerId}/sources:
    get:
      operationId: cabinetOfferSourcesList
      summary: List traffic sources for a joined offer
      parameters:
        - $ref: '#/components/parameters/OfferId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Source'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    post:
      operationId: cabinetOfferSourceCreate
      summary: Create a traffic source for a joined offer
      parameters:
        - $ref: '#/components/parameters/OfferId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /cabinet/offers/{offerId}/sources/{sourceId}:
    patch:
      operationId: cabinetOfferSourceUpdate
      summary: Update a traffic source
      parameters:
        - $ref: '#/components/parameters/OfferId'
        - $ref: '#/components/parameters/SourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
    delete:
      operationId: cabinetOfferSourceDelete
      summary: Delete a traffic source (only if it has no clicks)
      parameters:
        - $ref: '#/components/parameters/OfferId'
        - $ref: '#/components/parameters/SourceId'
      responses:
        '204':
          description: No content
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /cabinet/source-groups:
    get:
      operationId: cabinetSourceGroupsList
      summary: List partner's source groups
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SourceGroup'
        '401':
          $ref: '#/components/responses/Error'
    post:
      operationId: cabinetSourceGroupCreate
      summary: Create a source group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceGroupInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceGroup'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /cabinet/source-groups/{groupId}:
    patch:
      operationId: cabinetSourceGroupUpdate
      summary: Update a source group
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceGroupInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceGroup'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
    delete:
      operationId: cabinetSourceGroupDelete
      summary: Delete an empty source group
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '204':
          description: No content
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
...
components:
  parameters:
...
    OfferId:
      name: offerId
      in: path
      required: true
      schema:
        type: string
    SourceId:
      name: sourceId
      in: path
      required: true
      schema:
        type: string
    GroupId:
      name: groupId
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    SourceTotals:
      type: object
      properties:
        clicks:
          type: integer
          format: int64
        unique_clicks:
          type: integer
          format: int64
        registrations:
          type: integer
          format: int64
        first_payments:
          type: integer
          format: int64
        income_kopecks:
          type: integer
          format: int64
    Source:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        comment:
          type: string
          nullable: true
        group_id:
          type: string
          nullable: true
        group_name:
          type: string
          nullable: true
        is_default:
          type: boolean
        is_active:
          type: boolean
        url:
          type: string
        totals:
          $ref: '#/components/schemas/SourceTotals'
        totals_30d:
          $ref: '#/components/schemas/SourceTotals'
        created_at:
          type: string
          format: date-time
    SourceInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        code:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        group_id:
          type: string
          nullable: true
        is_default:
          type: boolean
    SourceUpdate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        code:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        group_id:
          type: string
          nullable: true
        is_active:
          type: boolean
        is_default:
          type: boolean
    SourceGroup:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        comment:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SourceGroupInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        comment:
          type: string
          nullable: true
...